Yves Sablonier:
I was working on a Aerospatiale AS332 for some months. Now I
reconfigured the FDM with help of HHS and Maik Justus. There is still
some work to do, of course. But the as332 has a model, a FDM and some
basic instruments/animation and is now ready for basic test/use.